+ Исправление ошибок линтера + Добавление комментариев к новым пакетам/структурам/методам
